requirements.txt
wheel
click
python-dotenv
alembic
Flask
Flask-Cors
flask-marshmallow
Flask-RESTful
marshmallow
marshmallow-sqlalchemy
SQLAlchemy
sqlalchemy-mixins
SQLAlchemy-Utils
psycopg2-binary
requests
gunicorn
bcrypt==3.2.0
PyJWT==2.0.1
pika>=1.2.0,<2
apispec==6.0.2
apispec[validation]==6.0.2
apispec[yaml]==6.0.2
apispec-webframeworks==0.5.2
